fix (FederatedLogger): moved to callbacks module and added on learners initialization by default
