+++ p2pfl/learning/callbacks/__init__.py
@@ -1,2 +1,20 @@
-from .pytorch import scaffold_callback as pytorch_scaffold_callback
-from .tensorflow import scaffold_callback as tensorflow_scaffold_callback
+#
+# This file is part of the federated_learning_p2p (p2pfl) distribution
+# (see https://github.com/pguijas/p2pfl).
+# Copyright (c) 2022 Pedro Guijas Bravo.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, version 3.
+#
+# This program is distributed in the hope that it will be useful, but
+# WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+# General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
+#
+
+from .pytorch import scaffold_callback as pytorch_scaffold_callback  # noqa: F401
+from .tensorflow import scaffold_callback as tensorflow_scaffold_callback  # noqa: F401
